Use args.size in preference to one? and many?
The Enumerable method one? behaves in a similar manner to the any?
method such that it checks if the enum has one element that is
truthy. This obviously causes problems when trying to set false/nil
values as args.one? will return false.

Although many? does behave differently to one? and any? in that it
depends on the actual number of elements in the enum its use of a
block argument causes a performance impact on Ruby 1.8.7.