Fixes fragment caching to work with latest digests
In the current version of Rails the cache_digests code has been changed
to move fragment_name_with_digest back to a private method and expose
another method, cache_fragment_name, in CacheHelper. This fixes JBuilder
to properly call cache_fragment_name on the @context if it exists. It
then falls back to fragment_name_with_digest and finally to
ActiveSupport::Cache.