Fix passing object with partials w/o locals (#378). (#435)
* Fix passing object with partials w/o locals (#378).

Previous behavior: Passing objects to partial fails if the below format is used:
```
json.comments @post.comments, partial: 'comments/comment', as: :comment, show: false
```

It succeeds if locals option is used:
```
json.comments @post.comments, partial: 'comments/comment', locals: { show: false }, as: :comment
```

Swapping to options is a change to revitalize the conversation. I'm not
sure it's a permanent solution.

* Add spec to verify failure of locals param.

This spec validates @yusufali2205's concerns that passing in locals
explicitly no longer works.

* Attempt to fix spec and allow locals param.

This is not a good solution because it puts all of options, related to
locals or not, into the locals section of options. Will want advice for
a better solution, but this does make the spec pass.

* Fix locals in a cleaner way.

Revert previous attempt which duplicated options in favor of a version
that handles setting up locals in the same spot other locals are set up.

This allows us to only use the parameters that aren't used by the base
options to pull in locals.

* Key in to existing tests.

Rather than add completely new partials to highlight the issue, tie in
to the existing simple partial which still reproduces the error. Adds
specs for all 4 combinations of with/without partial key and
with/without locals key.

Delete existing added specs.

* Move the override one step lower.

The previous location relied on calling `_render_partial_with_options`
but there was another case where we call
```
json.bars foo.bars, partial: 'bars', as: :bar, include_baz: true
```
where, because there were args to the partial, we end up hitting
`_set_inline_partial` which did not route through
`_render_explicit_partial`. Putting it in a common location allows it to
work for both.