Update comment in controller generator templates
In 2013 this comment was updated in
rails/rails@3f9baeb2ec08138a0da09870ae60fd6b8165c07f to say "Only allow
a trusted parameter "white list" through." but no one ever noticed that
it didn't update in an app that used jbuilder.

jbuilder uses the same generators as Rails and actually overrides them
when you do a scaffold generator with jbuilder installed. That means
that this comment had the old text since 2013. Oops.

In #33681 we also stopped using "whitelist" in Rails so I've updated
this text and the text in Rails to remove this reference. See
rails/rails#37390