Document testing configuration for RSpec users
Jbuilder documentation does not guide users to any
extra configuration to test the response body of a
template.

Searching for "Rspec" in the project, we see that
there are lots of closed issues related to this issue.

Although this extra configuration is not related to JBuilder,
I believe this is an important thing to have in the documentation.

Including a note about it in the documentation would
save users hours of trying to debug JBuilder,
and maintainer's hours that could be used on another
features.